requirements.txt
amqp==5.0.2
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.3.0
black==20.8b1
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
dataclasses==0.8
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.13.0
graphene-django-optimizer==0.7.1
graphene-federation==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==3.1.0
kombu==5.0.2
mypy-extensions==0.4.3
pathspec==0.8.1
promise==2.3
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
redis==3.5.3
regex==2020.11.13
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unidecode==1.1.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
